requirements.txt
# ============================================================
# OjolBoost MAMS — Dependencies
# Framework: Google ADK + Vertex AI + MCP
# ============================================================

# --- Core ADK Framework ---
google-adk>=0.3.0

# --- Google Cloud ---
google-cloud-bigquery>=3.11.0
google-cloud-secret-manager>=2.16.0
google-auth>=2.22.0
vertexai>=1.38.0


# --- MCP Protocol ---
mcp>=1.0.0

# --- Data Validation (shared/schemas.py) ---
pydantic>=2.5.0
pydantic-settings>=2.1.0

# --- HTTP & APIs ---
requests>=2.31.0
httpx>=0.25.0

# --- Environment ---
python-dotenv>=1.0.0

# --- Utilities ---
pytz>=2023.3

# --- Testing ---
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# --- Dev Tools ---
black>=23.0.0
isort>=5.12.0
